+++ auth/init.js
@@ -1,13 +1,13 @@
-var db = require("../models/account");
+var db = require("../models");
 
 module.exports = function(passport) {
   passport.serializeUser(function(user, done) {
-    console.log("serializeUser");
+    console.log("Serialize User");
     done(null, user.id);
   });
 
   passport.deserializeUser(function(id, done) {
-    console.log("deseralizeUser");
+    console.log("Deseralize User");
     db.Account.findOne({
       where: {
         accountId: id
