+++ routes/apiRoutes.js
@@ -1,12 +1,13 @@
-module.exports = function(app, passport) {
-  console.log("Seting up routes...")
+var passport = require("passport");
+
+module.exports = function(app) {
+  console.log("Seting up routes...");
   app.get(
     "/auth/facebook",
-   // passport.authenticate("facebook", // {
-      // authType: "rerequest",
-      // scope: "user_location"
-    // }
-    // ),
+    passport.authenticate("facebook", {
+      authType: "rerequest",
+      scope: "user_location"
+    }),
     // eslint-disable-next-line no-unused-vars
     (req, res) => {
       console.log("Sign in sucessful!");
