+++ routes/htmlRoutes.js
@@ -54,7 +54,7 @@ module.exports = function(app) {
   });
 
   // Render 404 page for any unmatched routes
-  app.get("*", function(req, res) {
+  /* app.get("*", function(req, res) {
     res.render("404");
-  });
+  }); */
 };
