+++ views/layouts/main.handlebars
@@ -15,7 +15,6 @@
   {{> footer}}
 
   <script src="https://code.jquery.com/jquery.js"></script>
-  <script src="https://cdnjs.com/libraries/awesomplete"></script>
 </body>
 
 </html>
